move argument and hold into arg module
