+++ ogma/src/eng/var.rs
@@ -92,17 +92,6 @@ impl Variable {
     }
 }
 
-impl Hold {
-    pub fn ty(&self) -> std::borrow::Cow<Type> {
-        use std::borrow::Cow::*;
-        match self {
-            Hold::Lit(v) => Owned(v.ty()),
-            Hold::Var(v) => Borrowed(v.ty()),
-            Hold::Expr(s) => Borrowed(s.out_ty()),
-        }
-    }
-}
-
 // ###### SEED VARS ############################################################
 #[derive(Default, Debug)]
 pub struct SeedVars(Vec<(Str, Variable)>);
